requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
anthropic==0.42.0
mcp==1.26.0
anyio
python-dotenv==1.0.1

# SAML IdP — XML generation, signing, and X.509 cert management
lxml>=5.3.0
signxml>=3.2.0
cryptography>=43.0.0

# Dev / test dependencies
pytest
pytest-asyncio
